[WCNCR00148784] Fix coverity issue
[Detail]
Fix printf arg type mismatch.

MTK-Commit-Id: 215eef09058702c042e6203ceae795dea1c5b388

Feature: Wi-Fi
CR-Id: WCNCR00148784
Reviewed-on: http://gerrit.mediatek.inc:8080/1309575
CheckPatch: Check Patch <srv_checkpatch@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity@mediatek.com

+++ os/linux/hif/common/hal_pdma.c
@@ -145,7 +145,7 @@ u_int8_t halVerifyChipID(IN struct ADAPTER *prAdapter)
 	HAL_MCR_RD(prAdapter, prBusInfo->top_cfg_base + TOP_HW_CONTROL, &u4CIR);
 
 	DBGLOG(INIT, INFO, "WCIR_CHIP_ID = 0x%x, chip_id = 0x%x\n",
-		(u4CIR & WCIR_CHIP_ID), prChipInfo->chip_id);
+	       (uint32_t)(u4CIR & WCIR_CHIP_ID), prChipInfo->chip_id);
 
 	if ((u4CIR & WCIR_CHIP_ID) != prChipInfo->chip_id)
 		return FALSE;
